% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava_funset.R
\name{ans.out}
\alias{ans.out}
\title{export answers}
\usage{
ans.out(x, file = ".ans.csv", open = TRUE)
}
\arguments{
\item{x}{object will be exported}

\item{file}{what to store answers}

\item{open}{whether to open \code{file} containing answers, \code{TRUE}:yes;\code{no}.}
}
\description{
export answers
}
