% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava_funset.R
\name{cdout}
\alias{cdout}
\title{open folder}
\usage{
cdout(wd = getwd())
}
\arguments{
\item{wd}{directory that to open, default is current directory.}
}
\description{
open folder like \code{stata}'s command \code{cdout}, it is a convenient command.
}
\examples{
cdout()
cdout("C:/Windows")
}
