% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava.R
\name{DateIndex}
\alias{DateIndex}
\title{return date index}
\usage{
DateIndex(start = numeric(), end = numeric(), frequency = 1,
  ts.eps = getOption("ts.eps"))
}
\arguments{
\item{start}{a numeric vector like \code{c(year, month)}.}

\item{end}{like \code{start}.}

\item{frequency}{1: annual; 4: seasonal; monthly: 12.}
}
\description{
return date index: can return year and monthly or seasonal information when 
assign start and end date.
}
\details{
reference \link{ts}.
}
\examples{
(lst = DateIndex(start = c(2001,9), end = c(2011,7), frequency = 12) )
}
