% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava_funset.R
\name{growth}
\alias{growth}
\title{calculate growth of a series}
\usage{
growth(x, diff_gap = 1, method = c("division", "difference", "auto"),
  symrate = FALSE, standard = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{diff_gap}{the gap of calculating rate, you can get year on year rate of a annual series by specifing \code{diff_gap} for 12.}

\item{method}{\code{division}: \eqn{(x_t-x_{t-1})/x_{t-1}}; \code{difference}: \eqn{x_t-x_{t-1}}; 
\code{auto}: select method automatically.}

\item{symrate}{symmetry rate if \code{TRUE}.}

\item{standard}{whether to standard the ratio. \code{FALSE}:no, \code{TRUE}:yes.}
}
\description{
calculate some type of growth such as rate of increase, symmetrical rate, in addition get chain relative ratio or 
other type ratios by specifying gap.
}
\examples{
x = 1:30
growth(x, method = "division" )
growth(x, method = 2)
growth(x, method = 3, symrate = T)
x[2] = -2
growth(x, method = 3, symrate = T)
}
