% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava_funset.R
\name{impute_es}
\alias{impute_es}
\title{imputed by HoltWinters method}
\usage{
impute_es(x, extent = numeric(), forward = FALSE, methods = "HoltWinters",
  alpha = NULL, beta = NULL, gamma = NULL, file = paste(getwd(),
  "/HoltWintersAns.csv", sep = ""), out = TRUE)
}
\arguments{
\item{x}{a vector, should converted to \code{ts} if not a annual time series.}

\item{extent}{specify how long to predict forward or afterward.}

\item{forward}{if \code{TRUE}, will predicts \code{extent} values forward, default is \code{FALSE}.}

\item{methods}{method for impute, now only \code{HoltWinters}.}

\item{alpha}{reference to \link{HoltWinters}.}

\item{beta}{ditto.}

\item{gamma}{ditto.}

\item{file}{where to store answers.}

\item{out}{export answers if \code{TRUE}.}
}
\description{
imputed by HoltWinters method, forward or afterward
}
\examples{
x = c(12575.57 ,13965.77 ,14410.16 ,15531.18 ,17099.35 ,18231.83 ,19078.41 ,20062.66 ,21442.13 ,22879.18 ,23913.76 ,24365.62 ,25466.83 ,26441.65 ,27755.76 ,28762.68 ,30047.31 ,31553.62 ,32929.04 ,34601.72 ,36432.51 ,37241.35 ,38113.89 ,39591.86 ,41838.46 ,44218.31 ,46351.67 ,47954.53 ,48302.28 ,46909.42 ,48309.53 ,49725.5 ,51385.49 ,52615.34 ,54360.5)
x.ts = ts(x)
impute_es(x.ts, 6, gamma = F, forward = T)
impute_es(x.ts, 6, gamma = F, forward = T, out = F)
}
