% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava.R
\name{lag}
\alias{lag}
\title{return lag series}
\usage{
lag(x, k = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{k}{how many times to lag}
}
\description{
a simple method to get lag series
}
\details{
returned series will have \code{k} \code{NA} in the head
}
