% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava.R
\name{modeval}
\alias{modeval}
\title{Model Evaluate}
\usage{
modeval(y, y_hat, method = c("all", "MRE", "MAE", "RMSE", "Theil"))
}
\arguments{
\item{y}{original series}

\item{y_hat}{fitted or forecasted series by certain model}

\item{method}{how to calculate the error of \emph{y} and \emph{y_hat}, default is \emph{all}, otherwise,
\emph{MRE}: mean relative error, \emph{MAE}: mean absolute error, \emph{RMSE}: Root Mean Square Error,
\emph{Theil}: Theil coefficient.}
}
\value{
err a vector store the error of model calculated by \emph{method}.
}
\description{
Evaluate models by calculate error of fitting.
}
\examples{
x<-c(1,3,NA,3,56,2,6)
y<-c(1.1,2.3,NA,2.7,53,2.5,6.8)
x;y
modeval(x,y)
modeval(x,y,"MRE")
modeval(x,y,"MAE")
modeval(x,y,method="RMSE")
modeval(x,y,method="Theil")

modeval(x,y,method="Theilxxx") # fail
modeval(x, y , method = "R")
modeval(x, y , 9) # fail
}
