% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava.R
\name{reglist}
\alias{reglist}
\title{return a pretty answer of regression}
\usage{
reglist(..., type = "text")
}
\arguments{
\item{...}{objects}

\item{type}{default is \code{text}}
}
\description{
return a pretty answer of regression
}
