% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava.R
\name{similarcoef}
\alias{similarcoef}
\title{similar coefficient of industry structure}
\usage{
similarcoef(x1, x2)
}
\arguments{
\item{x1, x2}{industry's proportion of region 1 and region 2.}
}
\description{
similar coefficient of industry structure
}
\examples{
x1 = c(1:10)
x2 = c(1:10)
similarcoef(x1, x2)

x1 = rep(0,10)
x2 = rep(1,10)
similarcoef(x1,x2)
}
