% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava.R
\name{sortcol}
\alias{sortcol}
\title{order columns based on row's weighing means}
\usage{
sortcol(x, sortcols = c(1:length(x)), weight = NULL, decreasing = TRUE,
  mean_row = TRUE)
}
\arguments{
\item{x}{data.frame.}

\item{sortcols}{columns for comparing.}

\item{weight}{}

\item{decreasing}{\code{TRUE} decreasing order, default.}

\item{mean_row}{whether to append a row including mean values to last row.}
}
\description{
order columns based on row's weighing means
}
\examples{
year = c(2001:2005)
A = c(8,9,10,12,10)
B = c(9,10,11,10,9)
C = c(13,14,16,17,12)
weight = rep(1,5)
dt = data.frame(year,A,B,C,weight)

sortcol(dt)
sortcol(dt,mean_row = F)
sortcol(dt,sortcols = c(5,4,3,2,1) , weight = "B" , decreasing = F)
sortcol(dt,sortcols = c(2,3,4), weight = "weight")
}
