% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava.R
\name{tcsi}
\alias{tcsi}
\title{return time series with seasonal character randomly}
\usage{
tcsi(n, frequency = 12, cyl = 60, intercept = 100, slope = 1,
  s = c(1:12), irr_adj = 0.1)
}
\arguments{
\item{n}{number of observation}

\item{frequency}{monthly: 12; quarter: 4; annual: 1.}

\item{cyl}{cycle, mean long of cycle.}

\item{intercept}{intercept of trend line.}

\item{slope}{slope of trend line.}

\item{s}{a vector specifing seasonal factor, containing weights for every month.}

\item{irr_adj}{adjust irregular factor.}
}
\description{
return time series with seasonal character randomly, including \code{TC},\code{S},\code{I}.
}
\examples{
 
plot(tcsi,type = "l")

ts.tcsi<-ts(tcsi,frequency = 12)

plot(decompose(ts.tcsi))

}
