% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava.R
\docType{package}
\name{vava}
\alias{vava}
\alias{vava-package}
\title{frequently-used functions for me}
\description{
Some frequency-used functions for me, package them taking convenience for me and for you.
}
\author{
Hejun Liu, Beijing \email{liuhejunlj@163.com}
}
