% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vava_fun_content.R
\name{vava}
\alias{vava}
\title{list frequently-used functions
\code{vava} lists the list of frequently-used functions in \code{vava} package (function's content).}
\usage{
vava()
}
\examples{
vava()
}
\author{
Hejun Liu, Beijing \email{liuhejunlj@163.com}
}
